requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
Flask-Login>=0.6.0
Werkzeug>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.2.0
openpyxl>=3.1.0
pypdf2>=3.0.0
email-validator>=2.1.0
gunicorn>=21.2.0
python-bcb==0.3.3

